+++ src/citra_qt/bootmanager.h
@@ -13,6 +13,7 @@
 #include "common/thread.h"
 #include "core/core.h"
 #include "core/frontend/emu_window.h"
+#include "video_core/rasterizer_interface.h"
 
 class QKeyEvent;
 class QScreen;
@@ -116,6 +117,8 @@ signals:
     void DebugModeLeft();
 
     void ErrorThrown(Core::System::ResultStatus, std::string);
+    void DiskShaderCacheLoadingProgress(VideoCore::LoadCallbackStage stage, std::size_t value,
+                                        std::size_t total);
 };
 
 class OpenGLWindow : public QWindow {