+++ src/citra_qt/configuration/configure_dialog.cpp
@@ -84,7 +84,7 @@ void ConfigureDialog::PopulateSelectionList() {
 
     const std::array<std::pair<QString, QList<QWidget*>>, 4> items{
         {{QStringLiteral("General"),
-          {ui->generalTab, ui->webTab, ui->debugTab, ui->uiTab, ui->versionsTab}},
+          {ui->generalTab, ui->webTab, ui->debugTab, ui->uiTab, ui->releasesTab}},
          {QStringLiteral("System"), {ui->systemTab, ui->audioTab, ui->cameraTab}},
          {QStringLiteral("Graphics"), {ui->enhancementsTab, ui->graphicsTab, ui->customLayoutTab}},
          {QStringLiteral("Controls"), {ui->inputTab, ui->hotkeysTab}}}};
@@ -116,7 +116,7 @@ void ConfigureDialog::UpdateVisibleTabs() {
         {ui->debugTab, QStringLiteral("Debug")},
         {ui->webTab, QStringLiteral("Web")},
         {ui->uiTab, QStringLiteral("UI")},
-        {ui->versionsTab, QStringLiteral("Versions")},
+        {ui->releasesTab, QStringLiteral("Releases")},
     };
 
     ui->tabWidget->clear();